missing files to fix #358
